// src/api/generated/model/withholdingTaxResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * TaxMate API
 * 프리랜서 세금 관리 서비스 API
 * OpenAPI spec version: 1.0.0
 */

export interface WithholdingTaxResponse {
  grossAmount: number;
  incomeTax: number;
  localTax: number;
  totalTax: number;
  netAmount: number;
}