// src/api/generated/model/incomeTaxResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * TaxMate API
 * 프리랜서 세금 관리 서비스 API
 * OpenAPI spec version: 1.0.0
 */

export interface IncomeTaxResponse {
  annualIncome: number;
  totalExpenses: number;
  incomeAmount: number;
  taxableIncome: number;
  calculatedTax: number;
  localIncomeTax: number;
  totalTax: number;
  alreadyPaidTax: number;
  estimatedRefundOrPayment: number;
  isRefund: boolean;
  appliedTaxRate: string;
  basicDeduction: number;
}